requirements.txt
langchain
sentence-transformers
pdfplumber
google-generativeai
python-dotenv
Flask
langchain-community